+++ src/Flow/queryTrafficCountsByAirspace.test.ts
@@ -1,23 +1,15 @@
-import { inspect } from 'util';
-import { makeFlowClient } from '..';
+import { AssertionError } from 'chai';
 import moment from 'moment';
+import { inspect } from 'util';
+import { describe, expect, test } from 'vitest';
+import { makeFlowClient } from '..';
 import b2bOptions from '../../tests/options';
-import { FlowService } from '.';
-import { JestAssertionError } from 'expect';
-jest.setTimeout(20000);
+import { shouldUseRealB2BConnection } from '../../tests/utils';
 
-const conditionalTest = (global as any).__DISABLE_B2B_CONNECTIONS__
-  ? test.skip
-  : test;
-const xconditionalTest = xtest;
+describe('queryTrafficCountsByAirspace', async () => {
+  const Flow = await makeFlowClient(b2bOptions);
 
-let Flow: FlowService;
-beforeAll(async () => {
-  Flow = await makeFlowClient(b2bOptions);
-});
-
-describe('queryTrafficCountsByAirspace', () => {
-  conditionalTest('LFEE5R, aggregated', async () => {
+  test.runIf(shouldUseRealB2BConnection)('LFEE5R, aggregated', async () => {
     try {
       const res = await Flow.queryTrafficCountsByAirspace({
         dataset: { type: 'OPERATIONAL' },
@@ -67,7 +59,7 @@ describe('queryTrafficCountsByAirspace', () => {
         });
       }
     } catch (err) {
-      if (err instanceof JestAssertionError) {
+      if (err instanceof AssertionError) {
         throw err;
       }