+++ vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    // NM B2B can be slow sometimes, set test timeout to 20s
+    testTimeout: 20000,
+  },
+});