+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 
+- Replace `jest` by `vitest`
 - Fix `RunwayConfiguration` typing (#106)
 
 # v0.6.1 (2023-06-01)